enable users to download hydro csv
The csv file is sent direct to the client without creating a temp file
in the server. I planned to create a checkbox for each scenario and
enable users to download all the data as a csv file but I found each
scenario can have different timestamps or even different length (e.g.
s1 is for one year and s2 is for ten years). It is not very meaningful
for me to combine all the scenarios together in a single csv file.